// src/api/activity/response/ActivityTypeDto.ts
export enum ActivityTypeDto {
    CREATE_USER= 'CREATE_USER',
    USER_LOG_IN_SUCCESS = 'USER_LOG_IN_SUCCESS',
    USER_LOG_IN_FAILED = 'USER_LOG_IN_FAILED',
    USER_LOG_OUT = 'USER_LOG_OUT',
    DELETE_USER = 'DELETE_USER',
    DOCUMENT_OPEN = 'DOCUMENT_OPEN',
    KEY_OPEN = 'KEY_OPEN',
    DOCUMENT_DOWNLOAD = 'DOCUMENT_DOWNLOAD',
    CREATE_PROJECT_MEMBER = 'CREATE_PROJECT_MEMBER',
    UPDATE_PROJECT_MEMBER = 'UPDATE_PROJECT_MEMBER',
    DELETE_PROJECT_MEMBER = 'DELETE_PROJECT_MEMBER',
    CREATE_PROJECT_ENVIRONMENT = 'CREATE_PROJECT_ENVIRONMENT',
    UPDATE_PROJECT_ENVIRONMENT = 'UPDATE_PROJECT_ENVIRONMENT',
    DELETE_PROJECT_ENVIRONMENT = 'DELETE_PROJECT_ENVIRONMENT',
    UPDATE_PROJECT = 'UPDATE_PROJECT',
    CREATE_RESOURCE = 'CREATE_RESOURCE',
    UPDATE_RESOURCE = 'UPDATE_RESOURCE',
    DELETE_RESOURCE = 'DELETE_RESOURCE'
}